2. Парковка машин [#853 #303187]

+++ src/main/java/ru/job4j/design/carparking/Car.java
@@ -2,7 +2,7 @@ package ru.job4j.design.carparking;
 
 public interface Car {
 
-    double getSize();
+    int getSize();
 
     String getName();
 }
