+++ src/main/java/ru/job4j/design/carparking/PassengerCar.java
@@ -5,7 +5,7 @@ import static ru.job4j.design.carparking.Constants.*;
 public class PassengerCar implements Car {
 
     private String name;
-    private final double size = PASSENGER_CAR_SIZE;
+    private final int size = PASSENGER_CAR_SIZE;
 
     public PassengerCar(String name) {
         this.name = name;
@@ -17,7 +17,7 @@ public class PassengerCar implements Car {
 
 
     @Override
-     public double getSize() {
+     public int getSize() {
         return size;
     }
